Add initial "%writeandexecute" magic
This magic writes the cell to a file and then executes the cell.
It can be used to export code to a python file and enable code reuse.

The magic was copied as is from
https://github.com/minrk/ipython_extensions/blob/master/extensions/writeandexecute.py

It's licensed under the "3-Clause (aka New) BSD". The above repo is
"Copyright © Min Ragan-Kelley. All Rights Reserved".