Update testing
The latest IPython conda package does not include "iptest", so
include a small wrapper. Call as "python test.py"...

Remove coverage from travis, as the way we test magics seem to
prevent that coverage "sees" this usage and therefore does not
give any coverage data for ipyext :-(

Also restructure the travis tests to work with the new testing
script -> no need to switch to a temp dir. Unfortunately, we
still need to install the package to make it findable by nose.
Not sure why :-(